just small fixes in gitignore and remove npm lock files
