+++ components/package.json
@@ -1,6 +1,9 @@
 {
-  "version": "0.1.0",
+  "name": "components",
+  "author": "nikita medvedev",
+  "version": "0.2.0",
   "license": "MIT",
+  "module": "dist/components.esm.js",
   "main": "dist/index.js",
   "typings": "dist/index.d.ts",
   "files": [
@@ -8,7 +11,7 @@
     "src"
   ],
   "engines": {
-    "node": ">=10"
+    "node": ">=12"
   },
   "scripts": {
     "start": "tsdx watch",
@@ -29,9 +32,6 @@
     "singleQuote": true,
     "trailingComma": "es5"
   },
-  "name": "components",
-  "author": "nikita medvedev",
-  "module": "dist/components.esm.js",
   "devDependencies": {
     "husky": "^4.3.0",
     "tsdx": "^0.13.3",
