+++ components/src/sum.ts
@@ -1,0 +1,6 @@
+export const sum = (a: number, b: number) => {
+  if ('development' === process.env.NODE_ENV) {
+    console.log('boop');
+  }
+  return a + b;
+};
