+++ webapp/package.json
@@ -1,20 +1,19 @@
 {
   "name": "webapp",
   "version": "0.1.0",
-  "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "@types/jest": "^24.9.1",
-    "@types/node": "^12.12.56",
+    "@types/node": "^12.12.57",
     "@types/react": "^16.9.49",
     "@types/react-dom": "^16.9.8",
+    "components": "0.2.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-scripts": "3.4.3",
-    "typescript": "^3.7.5",
-    "components": "0.1.0"
+    "typescript": "^4.0.2"
   },
   "scripts": {
     "start": "react-scripts start",
